May 20/code_debugger/requirements.txt
# AutoGen framework (official PyPI version)
pyautogen

# Gemini API client
google-generativeai

# Async support
aiofiles
asyncio

# Linting
pylint

streamlit
